Move ProfileBuffer & AuthInfo to public lib
