+++ windows_integration/public/src/lib.rs
@@ -1,4 +1,5 @@
 use kanidm_proto::v1::UnixUserToken;
+use windows::Win32::Foundation::UNICODE_STRING;
 
 pub enum AccountType {
     Person,
@@ -28,4 +29,14 @@ pub struct AuthenticateAccountRequest {
 pub struct AuthenticateAccountResponse {
     pub status: Result<(), AuthPkgError>,
     pub token: Option<UnixUserToken>,
+}
+
+// Authentication Package Logon User
+pub struct AuthInfo {
+    pub username: UNICODE_STRING,
+    pub password: UNICODE_STRING,
+}
+
+pub struct ProfileBuffer {
+    pub token: UnixUserToken,
 }
